vp9_decodeframe: promote row-mt calcs to size_t
This avoids integer overflow when decoding very large frames. `size_t`
is acceptable for 32-bit as a check in `vp9_dec_alloc_row_mt_mem()`,
which was added in:
  d9a548c01 vp9: fix for out-of-bounds write in decode coeffs
would have already failed. The calculations fixed in this change are
similar to the ones in the earlier change.

Bug: 496830326

+++ vp9/decoder/vp9_decodeframe.c
@@ -1804,7 +1804,8 @@ static void recon_tile_row(TileWorkerData *tile_data, VP9Decoder *pbi,
   for (mi_col = mi_col_start; mi_col < mi_col_end; mi_col += MI_BLOCK_SIZE) {
     const int c = mi_col >> MI_BLOCK_SIZE_LOG2;
     int plane;
-    const int sb_num = (cur_sb_row * (aligned_cols >> MI_BLOCK_SIZE_LOG2) + c);
+    const size_t sb_num =
+        (cur_sb_row * (aligned_cols >> MI_BLOCK_SIZE_LOG2) + c);
 
     // Top Dependency
     if (cur_sb_row) {
@@ -1879,7 +1880,7 @@ static void parse_tile_row(TileWorkerData *tile_data, VP9Decoder *pbi,
     const int r = mi_row >> MI_BLOCK_SIZE_LOG2;
     const int c = mi_col >> MI_BLOCK_SIZE_LOG2;
     int plane;
-    const int sb_num = (r * (aligned_cols >> MI_BLOCK_SIZE_LOG2) + c);
+    const size_t sb_num = (r * (aligned_cols >> MI_BLOCK_SIZE_LOG2) + c);
     for (plane = 0; plane < MAX_MB_PLANE; ++plane) {
       tile_data->xd.plane[plane].eob =
           row_mt_worker_data->eob[plane] + (sb_num << EOBS_PER_SB_LOG2);
